Encode JSON payloads with serde
At this point the cast client can successfully connect to a chromecast,
establish a connection, and get the device's status. This is enough for
my Google Home to make the 'a device has connected' noise. \o/

Only one status message succeeds because the cast client uses a constant
request id.

The decoder has changed to produce decoded payloads (instances of payload::Payload).